# contract/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "RepoPilot Frontend–Backend API",
    "description": "Contract of HTTP APIs that the frontend (`frontend/src/lib/api.ts`, `auth.ts`) calls against the backend. Derived from frontend clients and response TypeScript types.",
    "version": "4.2.0"
  },
  "servers": [
    {
      "url": "http://localhost:8000",
      "description": "Local backend (dev). Vite proxies `/api` and `/auth` from :5173 to this origin."
    }
  ],
...
  "paths": {
...
    "/api/repos/{repoId}/knowledge/policy": {
      "get": {
        "tags": [
          "Knowledge"
        ],
        "summary": "Get knowledge-indexing policy",
        "operationId": "fetchKnowledgePolicy",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "$ref": "#/components/parameters/RepoId"
          }
        ],
        "responses": {
          "200": {
            "description": "Indexing policy and limits",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/KnowledgePolicy"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          }
        }
      }
    },
    "/api/repos/{repoId}/knowledge/build": {
      "post": {
        "tags": [
          "Knowledge"
        ],
        "summary": "Build / rebuild knowledge index for a repo",
        "operationId": "buildKnowledge",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "$ref": "#/components/parameters/RepoId"
          }
        ],
        "requestBody": {
          "required": false,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/KnowledgeBuildRequest"
              }
            }
          }
        },
        "responses": {
          "202": {
            "description": "Build task accepted",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/KnowledgeBuildResult"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "500": {
            "$ref": "#/components/responses/ServerError"
          }
        }
      }
    },
    "/api/repos/{repoId}/knowledge": {
      "get": {
        "tags": [
          "Knowledge"
        ],
        "summary": "Get knowledge overview for a repo",
        "operationId": "fetchKnowledge",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "$ref": "#/components/parameters/RepoId"
          },
          {
            "name": "commit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            },
            "deprecated": true,
            "description": "Deprecated compatibility parameter. CodeWiki GraphRAG represents current HEAD; historical commits remain available through JGit history and compare APIs."
          }
        ],
        "responses": {
          "200": {
            "description": "Knowledge overview",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/KnowledgeOverview"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "500": {
            "$ref": "#/components/responses/ServerError"
          }
        }
      }
    },
    "/api/repos/{repoId}/knowledge/commits": {
      "get": {
        "tags": [
          "Knowledge"
        ],
        "summary": "List Git commit history for the indexed repository",
        "description": "Returns JGit history from the local repository mirror. Only current HEAD has a CodeWiki GraphRAG index.",
        "operationId": "fetchKnowledgeCommits",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "$ref": "#/components/parameters/RepoId"
          }
        ],
        "responses": {
          "200": {
            "description": "Indexed commit list",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IndexedCommitListResponse"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          }
        }
      }
    },
    "/api/repos/{repoId}/knowledge/compare": {
      "get": {
        "tags": [
          "Knowledge"
        ],
        "summary": "Compare two indexed commits",
        "operationId": "compareKnowledgeCommits",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "$ref": "#/components/parameters/RepoId"
          },
          {
            "name": "base",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "Base commit SHA"
          },
          {
            "name": "head",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "Head commit SHA"
          }
        ],
        "responses": {
          "200": {
            "description": "Compare result",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CommitCompareResult"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          }
        }
      }
    },
    "/api/repos/{repoId}/knowledge/settings": {
      "put": {
        "tags": [
          "Knowledge"
        ],
        "summary": "Update knowledge index settings",
        "operationId": "updateKnowledgeSettings",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "$ref": "#/components/parameters/RepoId"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/KnowledgeSettingsRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Updated settings",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/KnowledgeSettings"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          }
        }
      }
    },
...
    "/api/repos/{repoId}/knowledge/tasks": {
      "get": {
        "tags": [
          "Knowledge"
        ],
        "summary": "List persisted knowledge build tasks",
        "operationId": "fetchKnowledgeBuildTasks",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "$ref": "#/components/parameters/RepoId"
          },
          {
            "name": "limit",
            "in": "query",
            "schema": {
              "type": "integer",
              "minimum": 1,
              "maximum": 100,
              "default": 20
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Build task history",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "items"
                  ],
                  "properties": {
                    "items": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/KnowledgeBuildTask"
                      }
                    }
                  }
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          }
        }
      }
    },
    "/api/repos/{repoId}/knowledge/tasks/{taskId}": {
      "get": {
        "tags": [
          "Knowledge"
        ],
        "summary": "Get a persisted knowledge build task",
        "operationId": "fetchKnowledgeBuildTask",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "$ref": "#/components/parameters/RepoId"
          },
          {
            "$ref": "#/components/parameters/TaskId"
          }
        ],
        "responses": {
          "200": {
            "description": "Build task",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/KnowledgeBuildTask"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          }
        }
      }
    },
    "/api/repos/{repoId}/knowledge/tasks/{taskId}/errors": {
      "get": {
        "tags": [
          "Knowledge"
        ],
        "summary": "List detailed errors and warnings for a build task",
        "operationId": "fetchKnowledgeBuildErrors",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "$ref": "#/components/parameters/RepoId"
          },
          {
            "$ref": "#/components/parameters/TaskId"
          }
        ],
        "responses": {
          "200": {
            "description": "Build errors",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "items"
                  ],
                  "properties": {
                    "items": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/KnowledgeBuildError"
                      }
                    }
                  }
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          }
        }
      }
    },
    "/api/repos/{repoId}/knowledge/quality": {
      "get": {
        "tags": [
          "Knowledge"
        ],
        "summary": "Get the latest persisted index quality state",
        "operationId": "fetchKnowledgeQuality",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "$ref": "#/components/parameters/RepoId"
          }
        ],
        "responses": {
          "200": {
            "description": "Index quality",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/KnowledgeQuality"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          }
        }
      }
    },
    "/api/repos/{repoId}/knowledge/search": {
      "post": {
        "tags": [
          "Knowledge"
        ],
        "summary": "Search indexed chunks with vector or keyword retrieval",
        "operationId": "searchKnowledge",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "$ref": "#/components/parameters/RepoId"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/KnowledgeSearchRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Ranked knowledge chunks",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/KnowledgeSearchResponse"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          }
        }
      }
    },
    "/api/repos/{repoId}/knowledge/graph/status": {
      "get": {
        "tags": [
          "Knowledge"
        ],
        "summary": "Get CodeWiki GraphRAG index status",
        "operationId": "fetchKnowledgeGraphStatus",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "$ref": "#/components/parameters/RepoId"
          }
        ],
        "responses": {
          "200": {
            "description": "GraphRAG index status and entity counts",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/KnowledgeGraphStatus"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          }
        }
      }
    },
    "/api/repos/{repoId}/knowledge/graph/search": {
      "get": {
        "tags": [
          "Knowledge"
        ],
        "summary": "Search the CodeWiki knowledge graph",
        "operationId": "searchKnowledgeGraph",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "$ref": "#/components/parameters/RepoId"
          },
          {
            "$ref": "#/components/parameters/GraphQuery"
          },
          {
            "$ref": "#/components/parameters/GraphNodeType"
          },
          {
            "$ref": "#/components/parameters/GraphLanguage"
          },
          {
            "$ref": "#/components/parameters/GraphPath"
          },
          {
            "$ref": "#/components/parameters/GraphLimit"
          }
        ],
        "responses": {
          "200": {
            "description": "Ranked graph entities",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/KnowledgeGraphSearchResponse"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          }
        }
      }
    },
...
    "/api/repos/{repoId}/knowledge/wiki/generate": {
      "post": {
        "tags": [
          "Knowledge"
        ],
        "summary": "Generate or regenerate the CodeWiki wiki on demand",
        "operationId": "generateKnowledgeWiki",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "$ref": "#/components/parameters/RepoId"
          },
          {
            "$ref": "#/components/parameters/WikiLanguage"
          }
        ],
        "responses": {
          "202": {
            "description": "Wiki generation accepted",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/KnowledgeWikiGeneration"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          }
        }
      }
    },
    "/api/repos/{repoId}/knowledge/wiki": {
      "get": {
        "tags": [
          "Knowledge"
        ],
        "summary": "Get a generated CodeWiki wiki",
        "operationId": "fetchKnowledgeWiki",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "$ref": "#/components/parameters/RepoId"
          },
          {
            "$ref": "#/components/parameters/WikiLanguage"
          }
        ],
        "responses": {
          "200": {
            "description": "Generated wiki pages",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/KnowledgeWiki"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "GitHub OAuth access token",
        "description": "Token from GitHub OAuth (`Authorization: Bearer <token>`). Frontend stores it in localStorage and sends it via `apiFetch` when present."
      }
    },
    "parameters": {
      "RepoId": {
        "name": "repoId",
        "in": "path",
        "required": true,
        "schema": {
          "type": "string"
        },
        "description": "GitHub repository numeric id (string in frontend paths)"
      },
      "TaskId": {
        "name": "taskId",
        "in": "path",
        "required": true,
        "schema": {
          "type": "string",
          "format": "uuid"
        },
        "description": "Persisted knowledge build task id"
      },
...
    },
    "responses": {
...
      "Unauthorized": {
        "description": "Missing or invalid Bearer token. Frontend clears auth and redirects to `/login`.",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorDetail"
            }
          }
        }
      },
...
      "ServerError": {
        "description": "Internal server error",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorDetail"
            }
          }
        }
      }
    },
    "schemas": {
...
      "KnowledgePolicy": {
        "type": "object",
        "required": [
          "required",
          "recommended",
          "excludedDirs",
          "storeOnly",
          "displayOnly",
          "featureMatrix",
          "limits"
        ],
        "properties": {
          "required": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "recommended": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "excludedDirs": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "storeOnly": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "displayOnly": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "featureMatrix": {
            "type": "object",
            "additionalProperties": {
              "type": "object",
              "properties": {
                "needs": {
                  "type": "array",
                  "items": {
                    "type": "string"
                  }
                },
                "not_needed": {
                  "type": "array",
                  "items": {
                    "type": "string"
                  }
                }
              }
            }
          },
          "limits": {
            "type": "object",
            "additionalProperties": {
              "type": "number"
            }
          },
          "dataSourceTiers": {
            "type": "object",
            "additionalProperties": true
          },
          "multiRepoPortfolio": {
            "type": "object",
            "additionalProperties": true
          }
        }
      },
      "KnowledgeBuildRequest": {
        "type": "object",
        "properties": {
          "indexEachCommit": {
            "type": "boolean",
            "default": false,
            "deprecated": true,
            "description": "Deprecated compatibility option. New builds index the current repository state."
          },
          "maxCommits": {
            "type": "integer",
            "minimum": 1,
            "maximum": 100,
            "default": 30,
            "deprecated": true,
            "description": "Deprecated compatibility limit for per-commit indexing."
          },
          "commitShas": {
            "type": [
              "array",
              "null"
            ],
            "items": {
              "type": "string"
            }
          }
        }
      },
...
      "KnowledgeBuildResult": {
        "type": "object",
        "required": [
          "taskId",
          "repoId",
          "status",
          "message",
          "async"
        ],
        "properties": {
          "taskId": {
            "type": "string",
            "format": "uuid"
          },
          "repoId": {
            "type": "string"
          },
          "status": {
            "type": "string",
            "enum": [
              "queued",
              "indexing"
            ]
          },
          "message": {
            "type": "string"
          },
          "async": {
            "type": "boolean",
            "const": true
          }
        }
      },
...
      "IndexedCommitListResponse": {
        "type": "object",
        "required": [
          "items"
        ],
        "properties": {
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/IndexedCommit"
            }
          }
        }
      },
      "KnowledgeSettings": {
        "type": "object",
        "required": [
          "indexEachCommit",
          "maxCommits",
          "activeCommitSha"
        ],
        "properties": {
          "indexEachCommit": {
            "type": "boolean",
            "deprecated": true
          },
          "maxCommits": {
            "type": "integer",
            "deprecated": true
          },
          "activeCommitSha": {
            "type": "string",
            "deprecated": true
          }
        }
      },
      "KnowledgeSettingsRequest": {
        "type": "object",
        "properties": {
          "indexEachCommit": {
            "type": [
              "boolean",
              "null"
            ],
            "deprecated": true
          },
          "maxCommits": {
            "type": [
              "integer",
              "null"
            ],
            "deprecated": true
          },
          "activeCommitSha": {
            "type": [
              "string",
              "null"
            ],
            "deprecated": true
          }
        }
      },
...
    }
  }
}